# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: YourTable
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /auth/facebook:
    # binds a127 app logic to a route
    x-swagger-router-controller: UserController
    post:
      description: Returns user's profile and auth token
      # used as the method name of the controller
      operationId: facebookLogin
      parameters:
        - name: authBody
          in: body
          description: facebook token from fb
          required: true
          schema:
            properties:
              access_token:
                type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/LoginResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /profile:
     # binds a127 app logic to a route
    x-swagger-router-controller: UserController
    get:
      description: Returns user's profile
      operationId: getMyProfile
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Profile"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /update/profile:
     # binds a127 app logic to a route
    x-swagger-router-controller: UserController
    put:
      description: Returns user's profile
      operationId: updateProfile
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
        - name: profileUpdate
          in: body
          required: true
          description: update user's profile
          schema:
            properties:
              displayName:
                type: string
              email:
                type: string
              phone:
                type: string
              avatar:
                type: string
      responses:
        "204":
          description: Success
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /create/event:
     # binds a127 app logic to a route
    x-swagger-router-controller: EventController
    post:
      description: Create event with authorized user
      operationId: createEvent
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
        - name: eventCreate
          in: body
          required: true
          description: Event boday
          schema:
            properties:
              yelpID:
                type: string
              title:
                type: string
              description:
                type: string
              purpose:
                type: string
              startTime:
                type: string
              minAge:
                type: number
              maxAge:
                type: number
              capacity:
                type: number
              gender: 
                type: string
              lng:
                type: number
              lat: 
                type: number
      responses:
        "204":
          description: Success
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /event/{page},{pageSize}:
      x-swagger-router-controller: EventController
      get:
        description: get a page of events
        operationId: getEvents
        parameters:
          - name: page
            type: number
            in: path
            required: true
          - name: pageSize
            type: number
            in: path
            required: true
            default: 10
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/EventListResponse"
          # responses may fall through to errors
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /event/{id}:
    x-swagger-router-controller: EventController
    get:
      description: get a single event by mongo id
      operationId: getOne
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Event"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /restaurant:
    x-swagger-router-controller: RestaurantController
    post:
      description: Uploads csv file of restaurant/promotion info, admin required.
      operationId: createRestaurant
      consumes:
        - multipart/form-data
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
        - name: uploadFile
          in: formData
          type: file
          required: true
      responses:
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  Profile:
    type: object
    properties:
      id:
        type: string
      facebookID:
        type: string
      firebaseID:
        type: string
      firstname:
        type: string
      displayName:
        type: string
      lastname:
        type: string
      gender: 
        type: string
      email:
        type: string
      birthday: 
        type: string
      avatar:
        type: string
      phone:
        type: string
      createdAt:
        type: string
      updatedAt:
        type: string
  EventListResponse:
    required:
      - events
    properties:
      events:
        type: array
        items:
          type: object
          properties:
            title:
              type: string
            purpose:
              type: string
            yelpID:
              type: string
            capacity:
              type: number
            attendeesCount:
              type: number
            startTime:
              type: string
            author:
              type: string
            avatar:
              type: string
  Event:
    type: object
    properties:
      id:
        type: string
      yelpID:
        type: string
      title: 
        type: string
      description:
        type: string
      purpose:
        type: string
      startTime:
        type: string
      minAge:
        type: number
      maxAge:
        type: number
      capacity:
        type: number
      gender:
        type: string
      attendees:
        type: array
        items:
          $ref: "#/definitions/Profile"
      author:
          $ref: "#/definitions/Profile"
      hasFinished:
        type: boolean
      lng: 
        type: number
      lat:
        type: number
      createdAt:
        type: string
      updatedAt:
        type: string
        
  LoginResponse:
      type: object
      properties:
        profile: 
          $ref: "#/definitions/Profile"
        isNewUser:
          type: boolean
        token:
          type: string
                
  ErrorResponse:
      type: string